chore: Clean out all styles
When I'm building new things I seem to get hung up on styling and bothered by it not looking as I
imagine it in the beginning. With this commit Iäm cleaning out practically all styling, except som
essential global styles, and will focus on getting all the functionality in place before getting
started on the looks of it all.